Extract shared per-file fan-out driver for CLI commands
cmd_apply, cmd_apply_channel, cmd_undo, cmd_delete_tags,
cmd_check_tags, and cmd_max_amplitude all repeated the same ~60-line
progress-bar / par_iter / ordered-stdout-flush / counter loop plus the
JSON-or-dry-run epilogue. Replace the six copies with
for_each_file(files, opts, per_file) and the finish_with_summary /
finish_without_summary epilogues in commands/utils.rs.

cmd_apply's TSV line (post-apply min/max re-analysis) is appended to
the per-file text inside the closure, preserving output order. The
only non-observable difference is that JSON records are now collected
in all output modes instead of only JSON mode (they were already
discarded otherwise).
