+++ src/commands/apply.rs
@@ -1,16 +1,13 @@
 use anyhow::Result;
 use colored::*;
 use mp3rgain::{analyze, steps_to_db, Channel};
-use rayon::prelude::*;
-use std::io::{self, Write};
+use std::fmt::Write as _;
 use std::path::PathBuf;
 
 use crate::cli::options::{Options, OutputFormat};
-use crate::commands::threading::effective_threads;
-use crate::commands::utils::{create_json_summary, print_dry_run_notice, update_counters};
-use crate::json_output::{JsonFileResult, JsonOutput};
+use crate::commands::utils::{create_json_summary, finish_with_summary, for_each_file};
+use crate::json_output::JsonOutput;
 use crate::processors::apply::{process_apply, process_apply_channel};
-use crate::progress::{create_progress_bar, progress_finish, progress_inc, progress_set_message};
 use crate::util::get_filename;
 
 pub fn cmd_apply(files: &[PathBuf], steps: i32, opts: &Options) -> Result<()> {
@@ -51,112 +48,26 @@ pub fn cmd_apply(files: &[PathBuf], steps: i32, opts: &Options) -> Result<()> {
         println!();
     }
 
-    let pb = create_progress_bar(files.len(), opts);
-    let mut json_results: Vec<JsonFileResult> = Vec::with_capacity(files.len());
-    let mut successful = 0;
-    let mut failed = 0;
-
-    let parallel = effective_threads(opts) > 1 && files.len() > 1;
-
-    if parallel {
-        let pb_ref = pb.as_ref();
-        let collected: Vec<(JsonFileResult, String, Option<String>)> = files
-            .par_iter()
-            .map(|file| -> Result<(JsonFileResult, String, Option<String>)> {
-                let (result, text) = process_apply(file, steps, opts)?;
-                let tsv_line = if opts.output_format == OutputFormat::Tsv {
-                    analyze(file).ok().map(|info| {
-                        format!(
-                            "{}\t{}\t{:.1}\t{:.6}\t{}\t{}",
-                            get_filename(file),
-                            steps,
-                            db_value,
-                            1.0,
-                            info.max_gain(),
-                            info.min_gain()
-                        )
-                    })
-                } else {
-                    None
-                };
-                if let Some(pb) = pb_ref {
-                    pb.set_message(get_filename(file).to_string());
-                    pb.inc(1);
-                }
-                Ok((result, text, tsv_line))
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        let stdout = io::stdout();
-        let mut handle = stdout.lock();
-        for (_, text, tsv) in &collected {
-            if !text.is_empty() {
-                handle.write_all(text.as_bytes())?;
-            }
-            if let Some(line) = tsv {
-                writeln!(handle, "{}", line)?;
+    let (json_results, successful, failed) = for_each_file(files, opts, |file| {
+        let (result, mut text) = process_apply(file, steps, opts)?;
+        if opts.output_format == OutputFormat::Tsv {
+            if let Ok(info) = analyze(file) {
+                writeln!(
+                    text,
+                    "{}\t{}\t{:.1}\t{:.6}\t{}\t{}",
+                    get_filename(file),
+                    steps,
+                    db_value,
+                    1.0,
+                    info.max_gain(),
+                    info.min_gain()
+                )?;
             }
         }
-        drop(handle);
+        Ok((Some(result), text))
+    })?;
 
-        for (result, _, _) in collected {
-            update_counters(&result, &mut successful, &mut failed);
-            if opts.output_format == OutputFormat::Json {
-                json_results.push(result);
-            }
-        }
-    } else {
-        for file in files {
-            let filename = get_filename(file);
-            progress_set_message(&pb, filename);
-
-            let (result, text) = process_apply(file, steps, opts)?;
-            if !text.is_empty() {
-                print!("{}", text);
-            }
-            update_counters(&result, &mut successful, &mut failed);
-
-            if opts.output_format == OutputFormat::Tsv {
-                if let Ok(info) = analyze(file) {
-                    println!(
-                        "{}\t{}\t{:.1}\t{:.6}\t{}\t{}",
-                        filename,
-                        steps,
-                        db_value,
-                        1.0,
-                        info.max_gain(),
-                        info.min_gain()
-                    );
-                }
-            }
-
-            if opts.output_format == OutputFormat::Json {
-                json_results.push(result);
-            }
-
-            progress_inc(&pb);
-        }
-    }
-
-    progress_finish(pb);
-
-    if opts.output_format == OutputFormat::Json {
-        let output = JsonOutput {
-            files: Some(json_results),
-            album: None,
-            summary: Some(create_json_summary(
-                files.len(),
-                successful,
-                failed,
-                opts.dry_run,
-            )),
-        };
-        println!("{}", serde_json::to_string_pretty(&output)?);
-    } else {
-        print_dry_run_notice(opts);
-    }
-
-    Ok(())
+    finish_with_summary(files.len(), json_results, successful, failed, opts)
 }
 
 pub fn cmd_apply_channel(
@@ -201,78 +112,10 @@ pub fn cmd_apply_channel(
         println!();
     }
 
-    let pb = create_progress_bar(files.len(), opts);
-    let mut json_results: Vec<JsonFileResult> = Vec::with_capacity(files.len());
-    let mut successful = 0;
-    let mut failed = 0;
+    let (json_results, successful, failed) = for_each_file(files, opts, |file| {
+        let (result, text) = process_apply_channel(file, channel, steps, opts)?;
+        Ok((Some(result), text))
+    })?;
 
-    let parallel = effective_threads(opts) > 1 && files.len() > 1;
-
-    if parallel {
-        let pb_ref = pb.as_ref();
-        let collected: Vec<(JsonFileResult, String)> = files
-            .par_iter()
-            .map(|file| -> Result<(JsonFileResult, String)> {
-                let r = process_apply_channel(file, channel, steps, opts)?;
-                if let Some(pb) = pb_ref {
-                    pb.set_message(get_filename(file).to_string());
-                    pb.inc(1);
-                }
-                Ok(r)
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        let stdout = io::stdout();
-        let mut handle = stdout.lock();
-        for (_, text) in &collected {
-            if !text.is_empty() {
-                handle.write_all(text.as_bytes())?;
-            }
-        }
-        drop(handle);
-
-        for (result, _) in collected {
-            update_counters(&result, &mut successful, &mut failed);
-            if opts.output_format == OutputFormat::Json {
-                json_results.push(result);
-            }
-        }
-    } else {
-        for file in files {
-            let filename = get_filename(file);
-            progress_set_message(&pb, filename);
-
-            let (result, text) = process_apply_channel(file, channel, steps, opts)?;
-            if !text.is_empty() {
-                print!("{}", text);
-            }
-            update_counters(&result, &mut successful, &mut failed);
-
-            if opts.output_format == OutputFormat::Json {
-                json_results.push(result);
-            }
-
-            progress_inc(&pb);
-        }
-    }
-
-    progress_finish(pb);
-
-    if opts.output_format == OutputFormat::Json {
-        let output = JsonOutput {
-            files: Some(json_results),
-            album: None,
-            summary: Some(create_json_summary(
-                files.len(),
-                successful,
-                failed,
-                opts.dry_run,
-            )),
-        };
-        println!("{}", serde_json::to_string_pretty(&output)?);
-    } else {
-        print_dry_run_notice(opts);
-    }
-
-    Ok(())
+    finish_with_summary(files.len(), json_results, successful, failed, opts)
 }
